backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["LKJ <lkj@certi.org.br>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
